% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filler.R
\name{PCA.Filler}
\alias{PCA.Filler}
\title{PCA.Filler}
\usage{
PCA.Filler(data.mat, num.of.pcs = 2)
}
\arguments{
\item{data.mat}{a numeric matrix or data frame. Long, wide and squared matrix supported. NAs supported.}

\item{num.of.pcs}{is an integer specifying the number of components to be used. Default 2.}
}
\description{
Takes a matrix with NAs and returns a matrix without NAs.
Takes data.mat and calculates and returns predicted data.mat. Uses by default a 2 component PCA model.
No centering or scaling. Supports NAs and similar. Rownames/colnames are retained.
Columns/rows with only NAs (or similar) are removed.
}
\examples{
data.mat <- airquality[, 1:4]
data.mat
PCA.Filler(data.mat)
}
\author{
Berntsson, Martin <martin.berntsson@eon.se> and de Saint-Aubain, Philip Anton <philip-anton.desaint-aubain@knowit.dk>
}
